requirements/requirements_openIA.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
openai
pypdf